Fix swig bindings
 - move dynet::detail to expr.h so swig bindings can use it
 - two small fixes (include nodes.h, drop default parameter)
