+++ dynet/expr.cc
@@ -279,27 +279,6 @@ Expression to_device(const Expression & x, Device *device) {
 // Functions with variable argument lengths   //
 ////////////////////////////////////////////////
 
-namespace detail {
-template <typename F, typename T>
-inline Expression f(const T& xs) {
-  DYNET_ARG_CHECK(xs.size() > 0, "Zero-size argument passed to function");
-  ComputationGraph *pg = xs.begin()->pg;
-  std::vector<VariableIndex> xis(xs.size());
-  int i = 0;
-  for (auto xi = xs.begin(); xi != xs.end(); ++xi) xis[i++] = xi->i;
-  return Expression(pg, pg->add_function<F>(xis));
-}
-template <typename F, typename T, typename T1>
-inline Expression f(const T& xs, const T1& arg1) {
-  DYNET_ARG_CHECK(xs.size() > 0, "Zero-size argument passed to function");
-  ComputationGraph *pg = xs.begin()->pg;
-  std::vector<VariableIndex> xis(xs.size());
-  int i = 0;
-  for (auto xi = xs.begin(); xi != xs.end(); ++xi) xis[i++] = xi->i;
-  return Expression(pg, pg->add_function<F>(xis, arg1));
-}
-} // namespace detail
-
 Expression affine_transform(const std::initializer_list<Expression> &xs) { return detail::f<AffineTransform>(xs); }
 Expression affine_transform(const std::vector<Expression> &xs) { return detail::f<AffineTransform>(xs); }
